ivs: move common definitions to new ivs module
